fix build doc 2
